reafctoring: api/user.rs (unpacking)
Also some trouble shooting of the previous commit.

+++ src/api/user.rs
@@ -93,15 +93,14 @@ pub async fn try_login(
     };
     let scylla_session = scylla_session!(session);
     let cache = cache!(shared_cache);
-    match db::users::get_user_password_hash(&scylla_session, username).await {
-        Some(secrets) => {
-            prelude::check_user_password(secrets, &req.username, &req.password, scylla_session, cache, new_token_holder).await
-        }
-        _ => {
-            logging::log("Failed because user password hash cannot be retrieved from scylla.", Some(function_name!()));
-            actix_web::HttpResponse::Unauthorized().body("Invalid username or password")
-        }
+    
+    if let Some(secrets) = db::users::get_user_password_hash(&scylla_session, username).await  {
+        // TODO: wow this returns a HTTP responder... why?
+        return prelude::check_user_password(secrets, &req.username, &req.password, scylla_session, cache, new_token_holder).await;
     }
+    
+    logging::log("Failed because user password hash cannot be retrieved from scylla.", Some(function_name!()));
+    return actix_web::HttpResponse::Unauthorized().body("Invalid username or password");
 }
 
 #[named]
@@ -111,6 +110,7 @@ pub async fn token_login(
     shared_cache: actix_web::web::Data<security::structures::MokaCache>,
     req: actix_web::web::Json<structures::TokenLoginUser>,
 ) -> impl actix_web::Responder {
+    
     let new_token_holder = structures::TokenHolder {
         token: security::token(),
     };
@@ -129,21 +129,18 @@ pub async fn token_login(
         Some(req.username.clone()),
     )
     .await
-    .is_some()
+    .is_none()
     {
-        match db::users::get_user_password_hash(&scylla_session, username).await {
-            Some(secrets) => {
-                prelude::check_user_password(secrets, &req.username, &req.password, scylla_session, cache, new_token_holder).await
-            }
-            _ => {
-                logging::log("Failed because user password hash cannot be retrieved from scylla.", Some(function_name!()));
-                actix_web::HttpResponse::Unauthorized().body("Invalid password")
-            }
-        }
-    } else {
         logging::log("Failed because user supplied token is incorrect.", Some(function_name!()));
-        actix_web::HttpResponse::Unauthorized().body("Invalid or expired token")
+        return actix_web::HttpResponse::Unauthorized().body("Invalid or expired token");
     }
+    if let Some(secrets) = db::users::get_user_password_hash(&scylla_session, username).await {
+        return prelude::check_user_password(secrets, &req.username, &req.password, scylla_session, cache, new_token_holder).await;
+    }
+    
+    logging::log("Failed because user password hash cannot be retrieved from scylla.", Some(function_name!()));
+    return actix_web::HttpResponse::Unauthorized().body("Invalid password");
+
 }
 
 #[named]
@@ -165,41 +162,38 @@ pub async fn get_user_servers(
         Some(req.username.clone()),
     )
     .await
-    .is_some()
+    .is_none()
     {
-        match db::server::fetch_user_servers(&scylla_session, req.username.clone()).await {
-            Some(sids) => {
-                let _ = db::prelude::insert_user_token(
-                    &scylla_session,
-                    &cache,
-                    db::structures::KeyUser {
-                        key: Some(security::armor_token(new_token_holder.token.clone())),
-                        username: Some(req.username.clone()),
-                    },
-                )
-                .await;
-
-                let _ = db::users::delete_token(
-                    &scylla_session,
-                    req.username.clone(),
-                    security::armor_token(req.token.clone()),
-                )
-                .await;
-
-                actix_web::HttpResponse::Ok().json(&structures::ServersList {
-                    token: new_token_holder.token.clone(),
-                    s_list: sids,
-                })
-            }
-            None => {
-                logging::log("no hash", Some(function_name!()));
-                actix_web::HttpResponse::NotFound().body("No servers found for user")
-            }
-        }
-    } else {
         logging::log("no token", Some(function_name!()));
-        actix_web::HttpResponse::Unauthorized().body("Invalid or expired token")
+        return actix_web::HttpResponse::Unauthorized().body("Invalid or expired token");
     }
+
+    if let Some(sids) = db::server::fetch_user_servers(&scylla_session, req.username.clone()).await {
+        let _ = db::prelude::insert_user_token(
+            &scylla_session,
+            &cache,
+            db::structures::KeyUser {
+                key: Some(security::armor_token(new_token_holder.token.clone())),
+                username: Some(req.username.clone()),
+            },
+        )
+        .await;
+
+        let _ = db::users::delete_token(
+            &scylla_session,
+            req.username.clone(),
+            security::armor_token(req.token.clone()),
+        )
+        .await;
+
+        return actix_web::HttpResponse::Ok().json(&structures::ServersList {
+            token: new_token_holder.token.clone(),
+            s_list: sids,
+        });
+    }
+    
+    logging::log("no hash", Some(function_name!()));
+    return actix_web::HttpResponse::NotFound().body("No servers found for user");
 }
 
 #[named]
@@ -221,38 +215,36 @@ pub async fn get_user_pfp(
         Some(req.username.clone()),
     )
     .await
-    .is_some()
+    .is_none()
     {
-        match db::users::fetch_user_pfp(&scylla_session, &req.username).await {
-            Some(pfp_row) => {
-                let _ = db::prelude::insert_user_token(
-                    &scylla_session,
-                    &cache,
-                    db::structures::KeyUser {
-                        key: Some(security::armor_token(new_token_holder.token.clone())),
-                        username: Some(req.username.clone()),
-                    },
-                )
-                .await;
-
-                let _ = db::users::delete_token(
-                    &scylla_session,
-                    req.username.clone(),
-                    security::armor_token(req.token.clone()),
-                )
-                .await;
-
-                actix_web::HttpResponse::Ok().json(&structures::GetUserPfpResp {
-                    token: new_token_holder.token.clone(),
-                    img_url: pfp_row.pfp,
-                })
-            }
-            None => actix_web::HttpResponse::NotFound().body("User not found."),
-        }
-    } else {
         logging::log("no token", Some(function_name!()));
-        actix_web::HttpResponse::Unauthorized().body("Invalid or expired token")
+        return actix_web::HttpResponse::Unauthorized().body("Invalid or expired token");
     }
+    
+    if let Some(pfp_row) = db::users::fetch_user_pfp(&scylla_session, &req.username).await {
+        let _ = db::prelude::insert_user_token(
+            &scylla_session,
+            &cache,
+            db::structures::KeyUser {
+                key: Some(security::armor_token(new_token_holder.token.clone())),
+                username: Some(req.username.clone()),
+            },
+        )
+        .await;
+
+        let _ = db::users::delete_token(
+            &scylla_session,
+            req.username.clone(),
+            security::armor_token(req.token.clone()),
+        )
+        .await;
+
+        return actix_web::HttpResponse::Ok().json(&structures::GetUserPfpResp {
+            token: new_token_holder.token.clone(),
+            img_url: pfp_row.pfp,
+        });
+    }
+    return actix_web::HttpResponse::NotFound().body("User not found.");
 }
 
 #[named]
@@ -274,44 +266,41 @@ pub async fn set_user_pfp(
         Some(req.username.clone()),
     )
     .await
-    .is_some()
+    .is_none()
     {
-        let img_opt = match req.img_url.as_deref() {
-            // If "" -> None
-            Some(s) if s.trim().is_empty() => None,
-            other => other,
-        };
-        match db::users::set_user_pfp(&scylla_session, &req.username, img_opt).await {
-            Some(Ok(())) => {
-                let _ = db::prelude::insert_user_token(
-                    &scylla_session,
-                    &cache,
-                    db::structures::KeyUser {
-                        key: Some(security::armor_token(new_token_holder.token.clone())),
-                        username: Some(req.username.clone()),
-                    },
-                )
-                .await;
-
-                let _ = db::users::delete_token(
-                    &scylla_session,
-                    req.username.clone(),
-                    security::armor_token(req.token.clone()),
-                )
-                .await;
-
-                actix_web::HttpResponse::Ok().json(&structures::GetUserPfpResp {
-                    token: new_token_holder.token.clone(),
-                    img_url: req.img_url.clone(),
-                })
-            }
-            Some(Err(_e)) => actix_web::HttpResponse::InternalServerError()
-                .body("Failed to update profile picture."),
-            None => actix_web::HttpResponse::InternalServerError()
-                .body("Failed to update profile picture."),
-        }
-    } else {
         logging::log("no token", Some(function_name!()));
-        actix_web::HttpResponse::Unauthorized().body("Invalid or expired token")
+        return actix_web::HttpResponse::Unauthorized().body("Invalid or expired token");
     }
+
+    let img_opt = match req.img_url.as_deref() {
+        Some(s) if s.trim().is_empty() => None,
+        other => other,
+    };
+
+    if db::users::set_user_pfp(&scylla_session, &req.username, img_opt).await.is_err() {
+        return actix_web::HttpResponse::InternalServerError()
+            .body("Failed to update profile picture.");
+    };
+
+    let _ = db::prelude::insert_user_token(
+        &scylla_session,
+        &cache,
+        db::structures::KeyUser {
+            key: Some(security::armor_token(new_token_holder.token.clone())),
+            username: Some(req.username.clone()),
+        },
+    )
+    .await;
+
+    let _ = db::users::delete_token(
+        &scylla_session,
+        req.username.clone(),
+        security::armor_token(req.token.clone()),
+    )
+    .await;
+
+    return actix_web::HttpResponse::Ok().json(&structures::GetUserPfpResp {
+        token: new_token_holder.token.clone(),
+        img_url: req.img_url.clone(),
+    });
 }
